deep delete of section and melete
git-svn-id: https://source.sakaiproject.org/contrib/foothill/melete/trunk@48291 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ melete-impl/src/java/org/sakaiproject/component/app/melete/SectionDB.java
@@ -25,6 +25,8 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.ListIterator;
+import java.util.Map;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -37,6 +39,8 @@ import org.hibernate.Transaction;
 import org.hibernate.exception.ConstraintViolationException;
 import org.sakaiproject.api.app.melete.MeleteCHService;
 import org.sakaiproject.api.app.melete.exception.MeleteException;
+import org.sakaiproject.entity.cover.EntityManager;
+import org.sakaiproject.entity.api.Reference;
 
 /**
  * @author Rashmi
@@ -52,14 +56,21 @@ import org.sakaiproject.api.app.melete.exception.MeleteException;
 public class SectionDB implements Serializable {
 	private HibernateUtil hibernateUtil;
 	private MeleteBookmarksDB bookmarksDB;
-	
+	private MeleteCHService meleteCHService;
+	private ModuleDB moduleDB;
+
+	public static final int MELETE_RESOURCE_ONLY=0;
+	public static final int SECTION_RESOURCE_ONLY=1;
+	public static final int MELETE_RESOURCE_SECTION_RESOURCE=2;
+	public static final int NONE_TO_DELETE=3;
+
 	 /** Dependency: a logger component. */
 	 private Log logger = LogFactory.getLog(SectionDB.class);
 
 	public SectionDB(){
 		hibernateUtil = getHibernateUtil();
 	}
-	
+
 	/**
 	 * Add section sets the not-null values not been populated yet and then
 	 * inserts the section into section table.
@@ -85,7 +96,7 @@ public class SectionDB implements Serializable {
 			  	// save object
 			  tx = session.beginTransaction();
 			  session.save(section);
-				
+
 			  if(!fromImport)
 			  {
 	//				 set xml structure for sequencing and placement of sections
@@ -95,7 +106,7 @@ public class SectionDB implements Serializable {
 			  SectionUtil.addSectiontoList(sectionsSeqXML, section.getSectionId().toString());
 			  sectionsSeqXML = SectionUtil.storeSubSections();
 			  module.setSeqXml(sectionsSeqXML);
-					
+
 			  session.saveOrUpdate(module);
 			  }
 			  tx.commit();
@@ -109,7 +120,7 @@ public class SectionDB implements Serializable {
 		     }
 			catch(ConstraintViolationException cve)
 			{
-				logger.error("constraint voilation exception" + cve.getConstraintName());	
+				logger.error("constraint voilation exception" + cve.getConstraintName());
 				throw cve;
 			}
 			catch(HibernateException he)
@@ -127,7 +138,7 @@ public class SectionDB implements Serializable {
 			}
 		return null;
 	}
-		
+
 	public Integer editSection( Section section) throws MeleteException
 	{
 		try{
@@ -180,13 +191,13 @@ public class SectionDB implements Serializable {
 				SectionResource secResource = (SectionResource)section.getSectionResource();
 				if(secResource == null)
 					secResource = new SectionResource();
-								
+
 				secResource.setSection(section);
 				secResource.setResource(melResource);
-								
+
 			  section.setModificationDate(new java.util.Date());
 			  section.setSectionResource(secResource);
-			
+
 		 // save object
 			  tx = session.beginTransaction();
 			  	  session.saveOrUpdate(melResource);
@@ -226,33 +237,74 @@ public class SectionDB implements Serializable {
 			}
 	}
 
-	 public void deleteSection(Section sec, String userId) throws MeleteException
-	 {
+	//Depending on the deleteFrom parameter, this method cleans out the section from various
+	//MELETE tables
+	private void deleteFromMeleteTables(Section sec, String userId, int deleteFrom, String embedResourceId) throws MeleteException
+	{
+		MeleteBookmarks mb = null;
+		SectionResource secRes = null;
 		 try{
 		       Transaction tx = null;
 		       Session session = hibernateUtil.currentSession();
-	 	
+
 		       try
 		       	{
 		    	   tx = session.beginTransaction();
-		    	   
-		    	   //Delete section
-		    	   sec.setDeleteFlag(true);
-		    	   session.saveOrUpdate(sec);
-		    	   
-		    	   Module module = (Module)sec.getModule();
-		    	   logger.debug("checking module element");
-		    	   if(module != null)
+
+		    	   secRes = (SectionResource) sec.getSectionResource();
+
+		    	   if (deleteFrom != NONE_TO_DELETE)
 		    	   {
+		    		 if ((deleteFrom == MELETE_RESOURCE_ONLY)||(deleteFrom == MELETE_RESOURCE_SECTION_RESOURCE))
+		    		 {
+		    		   MeleteResource melRes = null;
+                       //Delete from MELETE_RESOURCE table
+		    		   if (deleteFrom == MELETE_RESOURCE_ONLY) melRes = getMeleteResource(embedResourceId);
+		    		   if (deleteFrom == MELETE_RESOURCE_SECTION_RESOURCE) melRes = (MeleteResource) session.merge(secRes.getResource());
+	    	           session.delete(melRes);
+	    	          }
+
+		    		 if ((deleteFrom == SECTION_RESOURCE_ONLY)||(deleteFrom == MELETE_RESOURCE_SECTION_RESOURCE))
+		    		 {
+                       //Delete from SECTION_RESOURCE table
+		    	       secRes = (SectionResource) session.merge(secRes);
+		    	       session.delete(secRes);
+		    		 }
+		    	   }
+		    	   if (deleteFrom != MELETE_RESOURCE_ONLY)
+		    	   {
+		    	     Module module = (Module)sec.getModule();
+		    	     Integer sectionId = sec.getSectionId();
+		    	     logger.debug("checking module element");
+		    	     if(module != null)
+		    	     {
 		    	   		String sectionsSeqXML = module.getSeqXml();
 		    	   		logger.debug("module is not null so changing seq"+ sectionsSeqXML);
 		    	   		SubSectionUtilImpl SectionUtil = new SubSectionUtilImpl();
-		    	   		logger.debug("deleting section id from xmllist" + sec.getSectionId().toString());
-		    	   		sectionsSeqXML =SectionUtil.deleteSection(sectionsSeqXML, sec.getSectionId().toString());
+		    	   		logger.debug("deleting section id from xmllist" + sectionId.toString());
+		    	   		sectionsSeqXML =SectionUtil.deleteSection(sectionsSeqXML, sectionId.toString());
 		    	   		module.setSeqXml(sectionsSeqXML);
+		    	   		module.getSections().remove(new Integer(sectionId));
+		    	   		module = (Module) session.merge(module);
 		    	   		session.saveOrUpdate(module);
+
+		    	     }
+		    	     if (userId != null)
+				     {
+				        mb = new MeleteBookmarks();
+					    mb.setUserId(userId);
+					    mb.setCourseId(sec.getModule().getCoursemodule().getCourseId());
+					    mb.setModuleId(sec.getModuleId());
+					    mb.setSectionId(sectionId);
+				     }
+		    	     //Delete section
+		    	     sec = (Section) session.merge(sec);
+		    	     session.delete(sec);
+
 		    	   }
+
 		    	   tx.commit();
+		    	   logger.debug("Deleted section from everywhere");
 
 		       	}
 		    catch (HibernateException he)
@@ -263,23 +315,20 @@ public class SectionDB implements Serializable {
 		    }
 		    catch (Exception e) {
 		      if (tx!=null) tx.rollback();
-		      logger.error(e.toString());		     
+		      logger.error(e.toString());
 		      throw e;
 		    }
 		    finally
 			{
 		      	hibernateUtil.closeSession();
     		  }
-		    if (userId != null)
-		    {	
-		    MeleteBookmarks mb = new MeleteBookmarks();
-			mb.setUserId(userId);
-			mb.setCourseId(sec.getModule().getCoursemodule().getCourseId());
-			mb.setModuleId(sec.getModuleId());
-			mb.setSectionId(sec.getSectionId());
-			bookmarksDB.deleteBookmark(mb);
+		    if (deleteFrom != MELETE_RESOURCE_ONLY)
+		    {
+		      if (userId != null)
+		      {
+		    	bookmarksDB.deleteBookmark(mb);
+		      }
 		    }
-			
 		}
 	      catch (Exception ex)
 		  {
@@ -287,7 +336,134 @@ public class SectionDB implements Serializable {
 			  ex.printStackTrace();
 			  throw new MeleteException("delete_module_fail");
 		  }
+	}
+
+	public void deleteSection(Section sec, String courseId, String userId) throws MeleteException
+	 {
+		if (sec.getContentType().equals("notype"))
+		{
+			deleteFromMeleteTables(sec, userId, NONE_TO_DELETE, null);
+		}
+
+		if ((sec.getContentType().equals("typeLink"))||(sec.getContentType().equals("typeUpload")))
+		{
+			boolean resourceInUse = false;
+			String resourceId = sec.getSectionResource().getResource().getResourceId();
+			//Check in SECTION_RESOURCE table
+			List srUseList = checkInSectionResources(resourceId, courseId);
+			if (srUseList != null)
+			{
+				//This means there is the reference for this section
+				//as well as another section
+				if (srUseList.size() > 1)
+				{
+					//Resource being used elsewhere
+					resourceInUse = true;
+				}
+			    else
+			    {
+			      //Checks all typeEditor sections for embedded media references
+			      //to this resource
+				  List resourceUseList = findResourceInUse(resourceId, courseId);
+				  //This means there is atleast one typeEditor section
+				  //with media embedded reference to this resource
+				  if ((resourceUseList != null)&&(resourceUseList.size() > 0))
+				  {
+					  resourceInUse = true;
+				  }
+			    }
+				//If resource is being referenced from elsewhere,
+				//only delete from MELETE_SECTION and SECTION_RESOURCE tables
+			    if (resourceInUse == true)
+			    {
+				  deleteFromMeleteTables(sec, userId, SECTION_RESOURCE_ONLY, null);
+			    }
+			    //If resource is not being referenced from anywhere else,
+			    //remove from CH, delete from MELETE_SECTION, SECTION_RESOURCE, and MELETE_RESOURCE tables
+			    if (resourceInUse == false)
+			    {
+			      try
+			      {
+				    meleteCHService.removeResource(resourceId);
+			      }
+			      catch (Exception e)
+				  {
+					e.printStackTrace();
+					logger.error("SectionDB -- deleteSection -- error in delete resource" + e.toString());
+				  }
+				  deleteFromMeleteTables(sec,userId, MELETE_RESOURCE_SECTION_RESOURCE, null);
+			    }
+			}
+		}//End typeLink and typeUpload
+
+		if (sec.getContentType().equals("typeEditor"))
+		{
+		  List<String> secEmbed = null;
+		  String resourceId = sec.getSectionResource().getResource().getResourceId();
+		  //Store all embedded media references in a list
+		  try
+		  {
+			//This method returns references only in meleteDocs
+			secEmbed = meleteCHService.findAllEmbeddedImages(resourceId);
+		  }
+		  catch(Exception e)
+		  {
+			e.printStackTrace();
+			logger.error("SectionDB -- deleteSection - findAllEmbeddedImages failed" );
+		  }
+		  //Remove main section.html resource from CH
+		  try
+	      {
+		    meleteCHService.removeResource(resourceId);
+	      }
+	      catch (Exception e)
+		  {
+			e.printStackTrace();
+			logger.error("SectionDB -- deleteSection -- error in delete resource" + e.toString());
+		  }
+	      //Delete references to the main section.html section in MELETE_SECTION, SECTION_RESOURCE
+	      //and MELETE_RESOURCE
+		  deleteFromMeleteTables(sec, userId, MELETE_RESOURCE_SECTION_RESOURCE, null);
+
+		  //Media reference processing
+		  if ((secEmbed != null)&&(secEmbed.size() > 0))
+		  {
+			  for (ListIterator<String> i = secEmbed.listIterator(); i.hasNext(); )
+			  {
+				  resourceId = (String)i.next();
+				  boolean mediaInUse = false;
+
+				  //Check to see if this media is also associated with a typeLink
+				  //or typeUpload section
+				  List srUseList = checkInSectionResources(resourceId, courseId);
+				  //If it is, continue to the next media reference
+				  if ((srUseList != null) &&(srUseList.size() > 0)) continue;
+
+				  //Check if this media is being referenced in any other typeEditor
+				  //section, since this main section has been deleted already it will
+				  //not check in its own section
+				  List resourceUseList = findResourceInUse(resourceId, courseId);
+				  //If it is, continue to the next media reference
+				  if ((resourceUseList != null) && (resourceUseList.size() > 0)) continue;
+
+				  //If it comes here, it means this media is only referenced within this section
+				  //so remove from CH and delete from MELETE_RESOURCE
+				  try
+			      {
+				    meleteCHService.removeResource(resourceId);
+			      }
+			      catch (Exception e)
+				  {
+					e.printStackTrace();
+					logger.error("SectionDB -- deleteSection -- error in delete resource" + e.toString());
+				  }
+			      deleteFromMeleteTables(sec, userId, MELETE_RESOURCE_ONLY, resourceId);
+			  }//End for loop
+		  }//End if secEmbed != null
+		}//end typeEditor
+
 	 }
+
 	 public Section getSection(int sectionId) throws HibernateException {
 		 	Section sec = null;
 		 	try
@@ -640,40 +816,88 @@ public class SectionDB implements Serializable {
 			ex.printStackTrace();
 			throw new MeleteException("add_section_fail");
 			}
-	}	
-	
-	public List<String> checkInSectionResources(String selResourceId, String courseId)
+	}
+
+	public List checkInSectionResources(String selResourceId, String courseId)
 	{
-		
+
 		try{
 		     Session session = hibernateUtil.currentSession();
-		     
-		     String queryString = "Select a from Section a, SectionResource b, CourseModule c where a.sectionId = b.sectionId AND " + 
-		     					   "a.moduleId = c.moduleId AND c.courseId=:courseId AND c.deleteFlag=0" + 
+
+		     String queryString = "Select a from Section a, SectionResource b, CourseModule c where a.sectionId = b.sectionId AND " +
+		     					   "a.moduleId = c.moduleId AND c.courseId=:courseId AND c.deleteFlag=0" +
 		     					   " AND b.resource.resourceId=:resourceId AND a.deleteFlag=0" ;
-		     
+
 		     Query query = session.createQuery(queryString);
 		     query.setParameter("courseId",courseId);
 		     query.setParameter("resourceId",selResourceId);
 		     List result_list = query.list();
 		     if (result_list == null) return null;
 		     logger.debug("result_list for sec resources " + result_list.size() + result_list.toString());
-		     List foundResources = new ArrayList<String> (0);
+		     return result_list;
+		     /*List foundResources = new ArrayList<String> (0);
 		     for(Iterator<Section> itr=result_list.listIterator(); itr.hasNext();)
 		     {
 		    	 Section sec = itr.next();
 		    	 String foundAt = sec.getModule().getTitle() +" >> " + sec.getTitle();
 		    	 foundResources.add(foundAt);
 		     }
-		     return foundResources;		    
+		     return foundResources;*/
 		}
 		catch(Exception ex){
 			logger.error(ex.toString());
 			return null;
 			}
-	
+
 	}
-	
+	 public List findResourceInUse(String selResourceId, String courseId)
+	  {
+		  try{
+			  List resourceUseList = null;
+
+			  logger.debug("now looking in embed data as section resources don't have it");
+			  String lookingFor = "/access/meleteDocs/content" + selResourceId;
+			  //find in embedded data
+			  long starttime = System.currentTimeMillis();
+			  //resourceUseList = new ArrayList<String>(0);
+			  List<Module> modList = moduleDB.getModules(courseId);
+			  Iterator<Module> i = modList.iterator();
+			  while (i.hasNext())
+			  {
+				Module mod = i.next();
+
+				String modSeqXml = mod.getSeqXml();
+				SubSectionUtilImpl SectionUtil = new SubSectionUtilImpl();
+				List<Element> allsec = SectionUtil.getAllSections(modSeqXml);
+				for (Iterator<Element> itr = allsec.iterator(); itr.hasNext();)
+				{
+					Section sec = getSection(Integer.parseInt(((Element) itr.next()).attributeValue("id")));
+					if (!sec.getContentType().equals("typeEditor")) continue;
+					List<String> secEmbed = meleteCHService.findAllEmbeddedImages(sec.getSectionResource().getResource().getResourceId());
+
+					if (secEmbed != null && secEmbed.contains(lookingFor)) {
+						/*String foundAt = sec.getModule().getTitle() + " >> " + sec.getTitle();
+						resourceUseList.add(foundAt);*/
+						resourceUseList.add(sec);
+						long endtime = System.currentTimeMillis();
+						logger.debug("found in " +(endtime - starttime));
+						return  resourceUseList;
+					}
+				}
+
+			}
+				long endtime = System.currentTimeMillis();
+
+				logger.debug("time to process all files to get all embedded data" +(endtime - starttime));
+			return null;
+
+		  }catch(Exception ex)
+			{
+				logger.error("SectionServiceImpl --find resource in use failed" );
+				return null;
+			}
+	  }
+
 	public void deleteResourceInUse(String delResourceId,String courseId) throws MeleteException
 	{
 		try
@@ -687,6 +911,7 @@ public class SectionDB implements Serializable {
 						+ "AND secResource.resource.resourceId=:resourceId";
 
 				tx = session.beginTransaction();
+
 				int delResources = session.createQuery(queryString).setString("courseId", courseId).setString("resourceId", delResourceId)
 						.executeUpdate();
 
@@ -720,9 +945,9 @@ public class SectionDB implements Serializable {
 			ex.printStackTrace();
 			throw new MeleteException("delete_resource_fail");
 		}
-	
+
 	}
-	
+
 	/**
 	 * @return Returns the hibernateUtil.
 	 */
@@ -736,16 +961,35 @@ public class SectionDB implements Serializable {
 	public void setHibernateUtil(HibernateUtil hibernateUtil) {
 		this.hibernateUtil = hibernateUtil;
 	}
-	
+
 	public void setBookmarksDB(MeleteBookmarksDB bookmarksDB)
 	{
 		this.bookmarksDB = bookmarksDB;
-	}	
+	}
 
 	/**
 	 * @param logger The logger to set.
 	 */
 	public void setLogger(Log logger) {
 		this.logger = logger;
-	}	
+	}
+
+	public MeleteCHService getMeleteCHService()
+	{
+		return this.meleteCHService;
+	}
+
+	public void setMeleteCHService(MeleteCHService meleteCHService)
+	{
+		this.meleteCHService = meleteCHService;
+	}
+	/**
+	 * @param moduleDB The moduleDB to set.
+	 */
+	public void setModuleDB(ModuleDB moduleDB) {
+		this.moduleDB = moduleDB;
+	}
+	public ModuleDB getModuleDB() {
+		return moduleDB;
+	}
 }
